+++ Assets/HUFEXT/GenericGDPR/CHANGELOG.md
@@ -1,3 +1,32 @@
+## [2.6.2] - 2021-01-27
+### Fixed
+- GDPR prefab no longer fullscreen
+
+
+## [2.6.1] - 2021-01-08
+### Fixed
+- Dependencies
+
+
+## [2.6.0] - 2020-12-23
+### Added
+- Header background to the popup
+
+### Fixed
+- Multiple warnings
+
+
+## [2.5.1] - 2020-12-02
+### Fixed
+- allow to accept only GDPR in example prefab
+
+
+## [2.5.0] - 2020-12-02
+### Added
+- GenericDialog dependency
+- API for handling additional policies(ATT and PersonalizedAds)
+
+
 ## [2.4.6] - 2020-09-17
 ### Added
 - Hufdefine file
@@ -13,33 +42,41 @@
 ### Fixed
 - Documentation errors
 
+
 ## [2.4.3] - 2020-08-17
 ### Added
 - Documentation
+
 
 ## [2.4.2] - 2020-07-29
 ### Changed
 - es texts update
 
+
 ## [2.4.1] - 2020-06-05
 ### Changed
 - One frame delay on AutoInit - fix wrong GDPR displayment on some devices
+
 
 ## [2.4.0] - 2020-06-01
 ### Added
 - Added optional feature to show GDPR only in specific country.
 
+
 ## [2.3.4] - 2020-05-21
 ### Fixed
 - Missing characters in Japanese and Korean font assets
+
 
 ## [2.3.3] - 2020-05-13
 ### Changed
 - Changed header text
 
+
 ## [2.3.2] - 2020-05-12
 ### Fixed
 - HIP initialization check differing from autoinit ones
+
 
 ## [2.3.1] - 2020-04-22
 ### Added
@@ -47,6 +84,7 @@
 
 ### Fixed
 - Fixed issue with translations on ios.
+
 
 ## [2.3.0] - 2020-04-15
 ### Added
@@ -56,6 +94,7 @@
 - Moved font assets to HUFEXT/GenericGDPR/Fonts directory.
 - Moved font configuration from GDPR config to GDPRView component in prefab.
 
+
 ## [2.2.0] - 2020-03-30
 ### Added
 - Added link for advertising network companies
@@ -64,9 +103,11 @@
 ### Changed
 - Dispose method will be called only if DestroyOnAccept flag is enabled
 
+
 ## [2.1.1] - 2020-03-30
 ### Changed
 - fixed namespaces
+
 
 ## [2.1.0] - 2020-03-20
 ### Added
@@ -78,9 +119,11 @@
 ### Changed
 - Small adjustments in GDPR prefab.
 
+
 ## [2.0.1] - 2020-03-18
 ### Fixed
 - GDPR window will not appear when personalized ads consent was revoked.
+
 
 ## [2.0.0] - 2020-02-26
 ### Added
@@ -90,6 +133,7 @@
 - Breaking code changes.
 - Updated policy text.
 - New UI design.
+
 
 ## [1.1.1] - 2020-01-08
 ### Added